cohorts/data-warehouse/pyproject.toml
[project]
name = "data-warehouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-storage>=3.10.0",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
]
